+++ changelog.md
@@ -9,6 +9,7 @@ ex: JF: added robot images to the assets folder
 ## [Unreleased]
 ### Added
 
+- AZ: Added Controller Robot.php, view robots.php, CSS doc assembly.css, and bots folder with pictures to the /data folder
 - MG: added handling for parts subpages (routing is currently limited to part a1 as demo)
 - MG: added parts model, parts view, and parts controller
 - AR: added history model
@@ -20,6 +21,7 @@ ex: JF: added robot images to the assets folder
 
 ### Changed
 
+- AZ: edited routes.php to route to assembly page. Added comments to Robots.php
 - MG: Parts controller updated to handle subpages, test routing done, comments added
 - MG: Updated css to improve parts view, updated model in prep for next issue.
 - AR: update history model to include all types in every sub array